// packages/core/src/API.ts
export type Channel = {
  __typename: "Channel";
  name: string;
  data: string;
};
// ...
export type Websocket_streamsSubscription = {
  websocket_streams?: {
    __typename: "Channel";
    name: string;
    data: string;
  } | null;
};